feat(components): migrate SearchInput to TypeScript
feature/typescript-input
