+++ src/components/SearchInput/SearchInput.docs.mdx
@@ -1,5 +1,5 @@
 import { Status, Props, Story } from '../../../.storybook/components';
-import SearchInput from './SearchInput';
+import { SearchInput } from './SearchInput';
 
 # SearchInput
 
@@ -9,7 +9,6 @@ Search Input is an input that indicates to the user they can use the field
 to search information on the screen.
 
 <Story id="forms-input-searchinput--base" />
-
 <Props of={SearchInput} />
 
 ## Component variations
