+++ src/components/SearchInput/index.ts
@@ -13,6 +13,6 @@
  * limitations under the License.
  */
 
-import SearchInput from './SearchInput';
+import { SearchInput } from './SearchInput';
 
 export default SearchInput;
